+++ src/lib/html.js
@@ -182,6 +182,10 @@ const { html, render } = (() => {
       }
     }
 
+    if (domNode.tagName.toLowerCase() === 'input') {
+      domNode.value = templateNode.value;
+    }
+
     if (domNode.tagName.indexOf('-') > -1 && templateNode.tagName.indexOf('-') > -1) {
       const templateSymbols = Object.getOwnPropertySymbols(templateNode);
       const domSymbols = Object.getOwnPropertySymbols(domNode);
